backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pydantic-settings==2.6.1
pydantic[email]==2.10.0
bcrypt==4.2.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
